syntax: handle UnaryExpr{Op:STAR, X:nil}
In "def f(*, x)", the unary expression has no operand.
Document this and be appropriately defensive.
Add another test (and example) of walk.

Also:
- visit Comprehension.Body before Clauses and fix existing TestWalk.
- move existing TestWalk

Fixes #157
