+++ syntax/walk.go
@@ -10,6 +10,9 @@ package syntax
 // recursively for each non-nil child of n.
 // Walk then calls f(nil).
 func Walk(n Node, f func(Node) bool) {
+	if n == nil {
+		panic("nil")
+	}
 	if !f(n) {
 		return
 	}
@@ -31,8 +34,8 @@ func Walk(n Node, f func(Node) bool) {
 		walkStmts(n.False, f)
 
 	case *AssignStmt:
-		Walk(n.RHS, f)
 		Walk(n.LHS, f)
+		Walk(n.RHS, f)
 
 	case *DefStmt:
 		Walk(n.Name, f)
@@ -97,10 +100,10 @@ func Walk(n Node, f func(Node) bool) {
 		}
 
 	case *Comprehension:
+		Walk(n.Body, f)
 		for _, clause := range n.Clauses {
 			Walk(clause, f)
 		}
-		Walk(n.Body, f)
 
 	case *IfClause:
 		Walk(n.Cond, f)
@@ -122,7 +125,9 @@ func Walk(n Node, f func(Node) bool) {
 		}
 
 	case *UnaryExpr:
-		Walk(n.X, f)
+		if n.X != nil {
+			Walk(n.X, f)
+		}
 
 	case *BinaryExpr:
 		Walk(n.X, f)
